+++ src/components/Contactpage/index.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const ContactPage = () => {
+  return (
+    <div>Contact Us</div>
+  )
+}
+
+export default ContactPage
