+++ src/layout/pages/Contact.jsx
@@ -1,8 +1,11 @@
 import React from 'react'
+import ContactPage from '../../components/Contactpage'
 
 const Contact = () => {
   return (
-    <div>Contact</div>
+ 
+      <ContactPage/>
+   
   )
 }
